+++ MCPluginMarketplace/src/main/java/com/danikvitek/MCPluginMarketplace/data/repository/CategoryRepository.java
@@ -2,11 +2,13 @@ package com.danikvitek.MCPluginMarketplace.data.repository;
 
 import com.danikvitek.MCPluginMarketplace.data.model.entity.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface CategoryRepository extends JpaRepository<Category, Integer> {
+public interface CategoryRepository extends JpaRepository<Category, Short> {
+//    @Query("select c from Category c where c.title = ?1")
     Optional<Category> findByTitle(String title);
 }
